refactor(oxlint-rules): move the guarded-endpoint factory to utils/
`rules/` holds twenty kebab-case rule modules; the shared helpers live in
`utils/` with camelCase names. The factory is a helper rather than a rule
module, so `utils/` is both its correct owning location and the directory
where its camelCase name matches its neighbours.

Addresses the review nit on #24535.
