+++ packages/twenty-oxlint-rules/rules/graphql-resolvers-should-be-guarded.ts
@@ -1,4 +1,4 @@
-import { createGuardedEndpointRule } from './createGuardedEndpointRule';
+import { createGuardedEndpointRule } from '../utils/createGuardedEndpointRule';
 
 export const RULE_NAME = 'graphql-resolvers-should-be-guarded';
