+++ packages/twenty-oxlint-rules/rules/rest-api-methods-should-be-guarded.ts
@@ -1,4 +1,4 @@
-import { createGuardedEndpointRule } from './createGuardedEndpointRule';
+import { createGuardedEndpointRule } from '../utils/createGuardedEndpointRule';
 
 export const RULE_NAME = 'rest-api-methods-should-be-guarded';
